cs-grader-server/requirements.txt
FastAPI==0.115.11
requests==2.32.3
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
cohere==4.51
pdf2image==1.16.3
Pillow>=10.0.0
python-multipart==0.0.9
google-cloud-vision==3.4.4
docker==7.1.0
colorama==0.4.6
pydantic-settings==2.2.1